+++ R/solve_POMDP.R
@@ -360,6 +360,9 @@ solve_POMDP <- function(model,
   if (is.character(model))
     model <- read_POMDP(model)
   
+  if (!inherits(model, "POMDP"))
+    stop("x needs to be a POMDP!") 
+  
   if (is.null(horizon))
     horizon <- model$horizon
   if (is.null(horizon))