+++ R/solve_SARSOP.R
@@ -92,6 +92,9 @@ solve_SARSOP <- function(model,
   if (is.character(model))
     model <- read_POMDP(model)
   
+  if (!inherits(model, "POMDP"))
+    stop("x needs to be a POMDP!") 
+  
   if (!is.null(model$horizon) &&
       !is.infinite(model$horizon))
     warning(