+++ tests/testthat/test-simulate.R
@@ -18,7 +18,7 @@ sim_r <- simulate_POMDP(Tiger, n = 1000, horizon = 5, return_beliefs = TRUE, ver
 expect_lt(abs(mean(sim_c$reward) - mean(sim_r$reward)) / mean(sim_c$reward), 0.2)
 
 ### sparse matrices
-problem <- read_POMDP(system.file("examples/shuttle.95.POMDP", package = "pomdp"), parse = TRUE)
+problem <- read_POMDP(system.file("examples/shuttle_95.POMDP", package = "pomdp"), parse = TRUE)
 sim_c <- simulate_POMDP(problem, n = 100, horizon = 10, verbose = verb)
 sim_r <- simulate_POMDP(problem, n = 100, horizon = 10, verbose = verb, engine = 'r')